Replace the color picker component with the new version
