+++ packages/components/src/color-picker/index.ts
@@ -1,0 +1,6 @@
+/**
+ * Internal dependencies
+ */
+import { LegacyAdapter as ColorPicker } from './legacy-adapter';
+
+export default ColorPicker;
