+++ packages/components/src/color-picker/legacy-adapter.tsx
