+++ packages/components/src/color-picker/stories/index.js
@@ -1,4 +1,10 @@
 /**
+ * External dependencies
+ */
+import { boolean, select } from '@storybook/addon-knobs';
+import colorize from 'tinycolor2';
+
+/**
  * WordPress dependencies
  */
 import { useState } from '@wordpress/element';
@@ -6,25 +12,70 @@ import { useState } from '@wordpress/element';
 /**
  * Internal dependencies
  */
-import ColorPicker from '../';
+import ColorPicker from '..';
+import { Flex } from '../../flex';
+import { Spacer } from '../../spacer';
+import { space } from '../../ui/utils/space';
 
-export default { title: 'Components/ColorPicker', component: ColorPicker };
+export default {
+	component: ColorPicker,
+	title: 'Components (Experimental)/ColorPicker',
+};
 
-const ColorPickerWithState = ( { ...props } ) => {
-	const [ color, setColor ] = useState( '#f00' );
+const PROP_UNSET = 'unset';
+
+const Example = () => {
+	const [ color, setColor ] = useState( undefined );
+	const props = {
+		enableAlpha: boolean( 'enableAlpha', false ),
+		copyFormat: select(
+			'copyFormat',
+			[ PROP_UNSET, 'rgb', 'hsl', 'hex' ],
+			PROP_UNSET
+		),
+	};
+
+	if ( props.copyFormat === PROP_UNSET ) {
+		delete props.copyFormat;
+	}
+
 	return (
-		<ColorPicker
-			{ ...props }
-			color={ color }
-			onChangeComplete={ ( value ) => setColor( value.hex ) }
-		/>
+		<Flex
+			as={ Spacer }
+			gap={ space( 2 ) }
+			justify="space-around"
+			align="flex-start"
+			marginTop={ space( 10 ) }
+		>
+			<ColorPicker { ...props } color={ color } onChange={ setColor } />
+			<div style={ { width: 200, textAlign: 'center' } }>
+				{ colorize( color ).toHslString() }
+			</div>
+			<ColorPicker { ...props } color={ color } onChange={ setColor } />
+		</Flex>
 	);
 };
 
 export const _default = () => {
-	return <ColorPickerWithState disableAlpha />;
+	return <Example />;
 };
 
-export const alphaEnabled = () => {
-	return <ColorPickerWithState disableAlpha={ false } />;
+const LegacyExample = () => {
+	const [ legacyColor, setLegacyColor ] = useState( '#fff' );
+	const legacyProps = {
+		color: legacyColor,
+		onChangeComplete: setLegacyColor,
+		disableAlpha: boolean( 'disableAlpha', true ),
+	};
+
+	return (
+		<Flex align="flex-start" justify="flex-start">
+			<ColorPicker { ...legacyProps } />
+			<pre style={ { width: '20em' } }>
+				{ JSON.stringify( legacyColor, undefined, 4 ) }
+			</pre>
+		</Flex>
+	);
 };
+
+export const legacy = () => <LegacyExample />;
