--- packages/components/src/ui/color-picker/index.ts
@@ -1,1 +1,0 @@
-export { LegacyAdapter as ColorPicker } from './legacy-adapter';
