+++ packages/e2e-tests/specs/editor/blocks/heading.test.js
@@ -11,8 +11,10 @@ import {
 describe( 'Heading', () => {
 	const CUSTOM_COLOR_TEXT = 'Custom color';
 	const CUSTOM_COLOR_BUTTON_X_SELECTOR = `//button[contains(text(),'${ CUSTOM_COLOR_TEXT }')]`;
+	const CUSTOM_COLOR_DETAILS_BUTTON_SELECTOR =
+		'.components-color-picker button[aria-label="Show detailed inputs"]';
 	const COLOR_INPUT_FIELD_SELECTOR =
-		'.components-color-palette__picker .components-text-control__input';
+		'.components-color-picker .components-input-control__input';
 	const COLOR_PANEL_TOGGLE_X_SELECTOR =
 		"//button[./span[contains(text(),'Color')]]";
 
@@ -80,10 +82,11 @@ describe( 'Heading', () => {
 		);
 
 		await customTextColorButton.click();
+		await page.click( CUSTOM_COLOR_DETAILS_BUTTON_SELECTOR );
 		await page.waitForSelector( COLOR_INPUT_FIELD_SELECTOR );
 		await page.click( COLOR_INPUT_FIELD_SELECTOR );
 		await pressKeyWithModifier( 'primary', 'A' );
-		await page.keyboard.type( '#7700ff' );
+		await page.keyboard.type( '7700ff' );
 		await page.click( 'h3[data-type="core/heading"]' );
 		await page.waitForSelector(
 			'.component-color-indicator[aria-label="(Color: #7700ff)"]'
